Show alternative taxonomies for lists templates
For example, in order to show the categories a post belongs to, the
desired taxonomy will be retrieved by a JSON file inside /data.

When listing categories, I want to show the tags each post inside a
category belongs to.

Similarly, when listing tags, I want the categories of each post.

But, when I'm listing technologies ("techs"), I just wanna show the
techs itself a project uses.