Add node script to send email with Buttondown API
This is how it works

We get two blog posts RSS feeds, one from the local build and one from
the production build.

The goals it to get the blog posts that are yet to be published, meaning
the posts not in the production RSS feed.

So we use the library rss-parser to parse the RSS feeds and get the
difference betwen the items.

Then we use Buttondown API to send the email.

The content of the email (body) is built with Hugo using a custom output
format named email.