Add option to use system preference to set theme
Now, when the user presses the theme button, it will open up a menu with
Light, Dark and Auto as option.

The auto option should respect the user's system preference. It is also
the default option if the user hasn't previously selected one.

This implementation was largely taken from w3c documents, like
https://www.w3.org/TR/wai-aria-practices/examples/menu-button/menu-button-actions.html.

A lot was done to make it accessible and keyboard friendly.