Fix infinite loop when styling navbar
...that occurred when a translateY is applied after the navbar
intersected with the main element, but then this translateY would
retrigger the intersection observer callback, which would cause an
infinite loop.

This StackOverflow question has some more details:
https://stackoverflow.com/questions/61699283/how-to-make-intersectionobserver-work-with-transform-translate-animations

I don't know if this solutions works, but it's at least not as easy to
reproduce the issue (navbar swinging up and down indefinitely).

The fix is to instead use a thin line (which is the top of the viewport)
as the container in the IntersectionObserver. Then, I check if the main
element and navbar intersected it. When they both do, it means the
navbar is inside the main element, then I hide the navbar with transform
etc.