Change completely how to do marquee effect
Just use a all JavaScript approach.

Without JavaScript, there will be a static image in the center of the
page.

But if JavaScript gets executed, we set up everything needed for the
marquee effect, which is to get the width of the image so that we can
calculate when the animation should reset.

Remove the img tag later