Simplify how to highlight current item in table of contents
The previous approach did not work, it was too buggy.

It handled H2 and other headings differently: if not H2, we tried to
highlight every sibling heading above it until we found an H2. This
proved to be more difficult than expected, so I removed it.

Now we only highlight one item at a time, because we reset every
highlighted item in the TOC before highlighting one.

This is the simplest approach, although it is possible that two heading
to be visible on the screen and it would be expected that the equivalent
TOC items to be highlighted, but it's not currently done.