Generate posts thumbnails for Open Graph and Twitter cards
This adds a script to generate thumbnails for each post.

It uses puppeteer to take the screenshot of a webpage.

The webpage is built as a custom output format named thumbnail.

There is also a custom output format named thumbails, to list every
thumbnail page for a section in a JSON, which we'll use in the script.

The script assumes a web server is running.