Ensure there is a unique image for each translated post
Before this, if a post was available in both portuguese and english,
the same image would be used for both of them, which is wrong.

Now, I added to the configuration a file name for each language.

I'm not taking advantage of the fact of Hugo's built-in Open Graph partial
template anymore, which is that it'll use an image with "cover",
"feature" or "thumbnail" as the post image.

Instead, I use the filename "post-image{,.pt}.png"

Because of this, I changed all instance of "thumbnail" to "post-image".